+++ SEQTaRget/R/SEQuential.R
@@ -7,7 +7,7 @@
 #' can be used to estimate observational analogs of intention-to-treat 
 #' and per-protocol effects, and can adjust for potential selection bias induced by losses-to-follow-up.
 #'
-#' @param data data.frame or data.table, will preform expansion according to arguments passed through the `options` argument
+#' @param data data.frame or data.table, will perform expansion according to arguments passed through the `options` argument
 #' @param id.col String: column name of the id column
 #' @param time.col String: column name of the time column
 #' @param eligible.col String: column name of the eligibility column
@@ -15,7 +15,7 @@
 #' @param outcome.col String: column name of the outcome column
 #' @param time_varying.cols List: column names for time varying columns
 #' @param fixed.cols List: column names for fixed columns
-#' @param method String: method of analysis to preform; should be one of `"ITT"`, `"dose-response"`, or `"censoring"`
+#' @param method String: method of analysis to perform; should be one of `"ITT"`, `"dose-response"`, or `"censoring"`
 #' @param options List: optional list of parameters from [SEQopts()]
 #' @param verbose Logical: if TRUE, cats progress to console, default is `TRUE`
 #'