+++ SEQTaRget/NEWS.md
@@ -1,3 +1,7 @@
+# SEQTaRget (development version)
+
+* Remove mention of units from time in docs.
+
 # SEQTaRget v1.4.1
 
 - Strip row-level vectors from fastglm objects to reduce weight.statistics memory usage and use a new internal function to print the coefficient table.