+++ SEQTaRget/NEWS.md
@@ -70,7 +70,7 @@
 
 # SEQTaRget v1.3.5
 
-- The `hazard_ratio()` function now correctly desscribes the estimate as "Hazard ratio"
+- The `hazard_ratio()` function now correctly describes the estimate as "Hazard ratio"
 - The bootstrapping now collects the log hazard ratio instead of the hazard ratio because the log hazard ratio has better normality properties.
 - The `covariates()` function now returns more nicely formatted output (with spaces around `~` and `+` symbols in the model formulae)