+++ SEQTaRget/NEWS.md
@@ -1,4 +1,4 @@
-# SEQTaRget (development version)
+# SEQTaRget v1.4.2
 
 * Remove mention of units from time in docs.
 * Improve memory usage in the bootstrapping.